check the arguments like printf

+++ example/test_plugin.c
@@ -34,7 +34,7 @@ static void test_handler(struct uh_connection *conn, int event)
     path = conn->get_path(conn);
 
     conn->send_head(conn, 200, -1, NULL);
-    conn->chunk_printf(conn, "Path: %.*s\n", path.len, path.p);
+    conn->chunk_printf(conn, "Path: %.*s\n", (int)path.len, path.p);
     conn->chunk_end(conn);
 }
